訂單功能 - 新增折扣活動

+++ springboot-mall/src/main/java/com/kuoning/springbootmall/dao/OrderDao.java
@@ -14,7 +14,7 @@ public interface OrderDao {
     Order getOrderById(Integer orderId);
 
     List<OrderItem> getOrderItemsById(Integer orderId);
-    Integer createOrder(Integer userId, int totalAmount);
+    Integer createOrder(Integer userId, double totalAmount);
 
     void createOrderItems(Integer orderId, List<OrderItem> orderItemList);
 }
